Module "usrmgr": Rework of the module
- Applied CGL
- Added CGL check to TravisCI
- Extracted classes to Classes/-Folder and applied namespace
- Extracted functions to Functions/-Folder
- Added PHPDocs
