+++ modules/usrmgr/Functions/FieldNeeded.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @param string $key
+ * @return int
+ */
+function FieldNeeded($key)
+{
+    global $cfg;
+
+    if ($cfg["signon_show_".$key] == 2) {
+        return 1;
+    } else {
+        return 0;
+    }
+}
