+++ modules/usrmgr/Functions/IfLowerOrEqualUserlevel.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @param int $userid
+ * @return bool
+ */
+function IfLowerOrEqualUserlevel($userid)
+{
+    global $line, $auth;
+
+    if ($line['type'] <= $auth['type']) {
+        return true;
+    } else {
+        return false;
+    }
+}
