+++ modules/usrmgr/Functions/IfLowerUserlevel.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @param int $userid
+ * @return bool
+ */
+function IfLowerUserlevel($userid)
+{
+    global $line, $auth;
+
+    if ($line['type'] < $auth['type']) {
+        return true;
+    } else {
+        return false;
+    }
+}
