+++ modules/usrmgr/Functions/UpdateUsrMgrUserFields.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * @param int $id
+ * @return bool
+ */
+function UpdateUsrMgrUserFields($id)
+{
+    global $db;
+
+    $db->qry("ALTER TABLE %prefix%user ADD %plain% VARCHAR(255) NOT NULL;", $_POST['name']);
+
+    return true;
+}
