+++ modules/usrmgr/add.php
@@ -1,357 +1,9 @@
 <?php
-include_once("modules/usrmgr/class_usrmgr.php");
-$usrmgr = new UsrMgr();
+
+$mail = new \LanSuite\Module\Mail\Mail();
+$usrmgr = new \LanSuite\Module\UsrMgr\UserManager($mail);
 
 $gd = new \LanSuite\GD();
-
-/**
- * @param int $id
- * @return bool
- */
-function UpdateUsrMgr($id)
-{
-    global $mf, $db, $usrmgr, $func, $cfg;
-
-    // Clan-Management
-    $clan = new \LanSuite\Module\ClanMgr\Clan();
-    if (ShowField('clan')) {
-        if ($_POST['new_clan_select']) {
-            $clan->Add($_POST['clan_new'], $id, $_POST["clanurl"], $_POST["newclanpw"]);
-        } elseif ($_POST['clan']) {
-            $clan->AddMember($_POST['clan'], $id);
-        } elseif (isset($_POST['clan'])) {
-            $clan->RemoveMember($id);
-        }
-    }
-
-    // Update User-Perissions
-    if ($id) {
-        $db->qry("DELETE FROM %prefix%user_permissions WHERE userid = %int%", $id);
-        if ($_POST["permissions"]) {
-            foreach ($_POST["permissions"] as $perm) {
-                $db->qry("INSERT INTO %prefix%user_permissions SET module = %string%, userid = %int%", $perm, $id);
-            }
-        }
-    }
-
-    // If new user has been added
-    if (!$mf->isChange) {
-        $usrmgr->WriteXMLStatFile();
-
-        // If auto generated PW, use PW stored in session, else use PW send by POST field
-        if ($_POST['password_original']) {
-            $_SESSION['tmp_pass'] = $_POST['password_original'];
-        }
-
-        if ($cfg["signon_password_mail"]) {
-            if ($usrmgr->SendSignonMail(0)) {
-                $func->confirmation(t('Dein Passwort und weitere Informationen wurden an deine angegebene E-Mail-Adresse gesendet.'), NO_LINK);
-            } elseif ($cfg['sys_internet']) {
-                $func->error(t('Es ist ein Fehler beim Versand der Informations-Email aufgetreten.') .'<br />'. t('Dein Passwort lautet: <b>%1</b>', array($_SESSION['tmp_pass'])), NO_LINK);
-            }
-        }
-
-        // Send email-verification link
-        if ($cfg['sys_login_verified_mail_only']) {
-            $usrmgr->SendVerificationEmail($id);
-        }
-
-        // Show passwort, if wanted, or has mail failed
-        if ($cfg['signon_password_view']) {
-            $func->information(t('Dein Passwort lautet: <b>%1</b>', array($_SESSION['tmp_pass'])), NO_LINK);
-        }
-        $_SESSION['tmp_pass'] = '';
-    }
-
-    return true;
-}
-
-/**
- * @param string $AvatarName
- * @return bool|string
- */
-function CheckAndResizeUploadPic($AvatarName)
-{
-    global $gd;
-
-    if ($AvatarName == '') {
-        return false;
-    }
-    $FileEnding = strtolower(substr($AvatarName, strrpos($AvatarName, '.'), 5));
-    if ($FileEnding != '.png' and $FileEnding != '.gif' and $FileEnding != '.jpg' and $FileEnding != '.jpeg') {
-        return t('Bitte eine Grafikdatei auswählen');
-    }
-
-    $gd->CreateThumb($AvatarName, $AvatarName, 100, 100);
-    return false;
-}
-
-/**
- * Check for optional gender selection
- *
- * @param int $gender   From Inputfield 0=None, 1=Male, 2=Female
- * @return bool|string  Returns Message on error else false
- */
-function check_opt_gender($gender)
-{
-    global $cfg;
-
-    if ($cfg["signon_show_gender"] == 2) {
-        if ($gender == 0) {
-            return t("Bitte wählen sie ein Geschlecht aus.");
-        } else {
-            return false;
-        }
-    }
-}
-
-/**
- * Check for optional birthday selection
- * If Date is (DateNow - 80 years) the Date is the present value.
- * From the display::AddDateTimeRow() function. Not the perfect way.
- *
- * @param string $date  From Inputfield like 2000-01-02
- * @return bool|string  Returns Message on error else false
- */
-function check_birthday($date)
-{
-    global $cfg;
-
-    if ($cfg["signon_show_birthday"] == 2) {
-        $ref_date = (date("Y")-80)."-".date("n")."-".date("d");
-        if ($date == $ref_date or ($date=="0000-00-00")) {
-            return t("Bitte das korrekte Geburtsdatum eingeben.");
-        } else {
-            return false;
-        }
-    }
-}
-
-/**
- * @param string $clanpw
- * @return bool|string
- */
-function CheckClanPWUsrMgr($clanpw)
-{
-    global $db, $auth;
-
-    if (!$_POST['new_clan_select'] and $auth['type'] <= 1 and $auth['clanid'] != $_POST['clan']) {
-        $clan = $db->qry_first("SELECT password FROM %prefix%clan WHERE clanid = %int%", $_POST['clan']);
-        if ($clan['password'] and $clan['password'] != md5($clanpw)) {
-            return t('Passwort falsch!');
-        }
-    }
-    return false;
-}
-
-/**
- * @param string $ClanName
- * @return bool|string
- */
-function CheckClanNotExists($ClanName)
-{
-    global $db;
-
-    $clan = $db->qry_first("SELECT 1 AS found FROM %prefix%clan WHERE name = %string%", $ClanName);
-    if ($clan['found']) {
-        return t('Dieser Clan existiert bereits!') .HTML_NEWLINE. t(' Wenn du diesem beitreten möchten, wähle ihn oberhalb aus dem Dropdownmenü aus.');
-    }
-
-    if (preg_match("/([.^\"\'`´]+)/", $ClanName)) {
-        return t('Du verwendest nicht zugelassene Sonderzeichen in deinem Clannamen.');
-    }
-
-    return false;
-}
-
-/**
- * @param string $field
- * @param int $mode
- * @param string $error
- * @return bool|string
- */
-function PersoInput($field, $mode, $error = '')
-{
-    global $dsp, $usrmgr, $smarty;
-
-    switch ($mode) {
-        case \LanSuite\MasterForm::OUTPUT_PROC:
-              $_POST[$field .'_1'] = substr($_POST[$field], 0, 11);
-              $_POST[$field .'_2'] = substr($_POST[$field], 13, 7);
-              $_POST[$field .'_3'] = substr($_POST[$field], 21, 7);
-              $_POST[$field .'_4'] = substr($_POST[$field], 35, 1);
-
-            if ($_POST[$field .'_1'] == '') {
-                $_POST[$field .'_1'] = "aaaaaaaaaaD";
-            }
-            if ($_POST[$field .'_2'] == '') {
-                  $_POST[$field .'_2'] = "bbbbbbb";
-            }
-            if ($_POST[$field .'_3'] == '') {
-                  $_POST[$field .'_3'] = "ccccccc";
-            }
-            if ($_POST[$field .'_4'] == '') {
-                  $_POST[$field .'_4'] = "d";
-            }
-
-              $smarty->assign('name', $field);
-              $smarty->assign('value1', $_POST[$field .'_1']);
-              $smarty->assign('value2', $_POST[$field .'_2']);
-              $smarty->assign('value3', $_POST[$field .'_3']);
-              $smarty->assign('value4', $_POST[$field .'_4']);
-            if ($error) {
-                  $smarty->assign('errortext', $dsp->errortext_prefix . $error . $dsp->errortext_suffix);
-            }
-            if (Optional("perso")) {
-                  $smarty->assign('optional', "_optional");
-            }
-
-            return $smarty->fetch('modules/usrmgr/templates/row_perso.htm');
-        break;
-
-        case \LanSuite\MasterForm::CHECK_ERROR_PROC:
-              $_POST[$field] = $_POST["perso_1"] . "<<" . $_POST["perso_2"] . "<". $_POST["perso_3"] . "<<<<<<<" . $_POST["perso_4"];
-            if ($_POST[$field] == "aaaaaaaaaaD<<bbbbbbb<ccccccc<<<<<<<d") {
-                $_POST[$field] = "";
-            }
-            if ($_POST[$field] == "<<<<<<<<<<") {
-                  $_POST[$field] = "";
-            }
-            if ($_POST[$field] != '') {
-                  $perso_res = $usrmgr->CheckPerso($_POST[$field]);
-                switch ($perso_res) {
-                    case 2:
-                        return str_replace("<", "&lt;", t('Das Format der Personalausweisnummer ist falsch. Bitte nach folgendem Muster eingeben: \'aaaaaaaaaaD<<bbbbbbb<ccccccc<<<<<<<d\''));
-                    break;
-                    case 3:
-                        return t('Prüfsummenfehler. Bitte überprüfen deine Angaben. Sehr wahrscheinlich hast du eine oder mehrere Zahlen falsch abgeschrieben.');
-                    break;
-                    case 4:
-                        return t('Dieser Personalausweis ist leider bereits abgelaufen.');
-                    break;
-                }
-            }
-            return false; // -> Means no error
-        break;
-    }
-}
-
-/**
- * @param string $field
- * @param int $mode
- * @param string $error
- * @return bool|string
- */
-function Addr1Input($field, $mode, $error = '')
-{
-    global $dsp;
-
-    switch ($mode) {
-        case \LanSuite\MasterForm::OUTPUT_PROC:
-            if ($_POST['street|hnr'] == '' and $_POST['street'] and $_POST['hnr']) {
-                $_POST['street|hnr'] = $_POST['street'] .' '. $_POST['hnr'];
-            }
-            $dsp->AddTextFieldRow('street|hnr', t('Straße und Hausnummer'), $_POST['street|hnr'], $error, '', Optional('street'));
-            return false;
-        break;
-
-        case \LanSuite\MasterForm::CHECK_ERROR_PROC:
-            if ($_POST['street|hnr'] != '' or FieldNeeded('street')) {
-                $pieces = explode(' ', $_POST['street|hnr']);
-                $_POST['hnr'] = array_pop($pieces);
-                $_POST['street'] = implode(' ', $pieces);
-
-                if ($_POST['street'] == '' or $_POST['hnr'] == '') {
-                    return t('Bitte gib Straße und Hausnummer in folgendem Format ein: "Straßenname 12".');
-                }
-            }
-            // Means no error
-            return false;
-        break;
-    }
-}
-
-/**
- * @param string $field
- * @param int $mode
- * @param string $error
- * @return bool|string
- */
-function Addr2Input($field, $mode, $error = '')
-{
-    global $dsp;
-
-    switch ($mode) {
-        case \LanSuite\MasterForm::OUTPUT_PROC:
-            if ($_POST['plz|city'] == '' and $_POST['plz'] and $_POST['city']) {
-                $_POST['plz|city'] = $_POST['plz'] .' '. $_POST['city'];
-            }
-            $dsp->AddTextFieldRow('plz|city', t('PLZ und Ort'), $_POST['plz|city'], $error, '', Optional('city'));
-            return false;
-        break;
-
-        case \LanSuite\MasterForm::CHECK_ERROR_PROC:
-            if (($_POST['plz|city'] != '') || (FieldNeeded('city'))) {
-                $pieces = explode(' ', $_POST['plz|city']);
-                $_POST['plz'] = array_shift($pieces);
-                $_POST['city'] = implode(' ', $pieces);
-
-                if ($_POST['plz'] == 0 or $_POST['city'] == '') {
-                    return t('Bitte gib Postleitzahl und Ort in folgendem Format ein: "12345 Stadt".');
-                } elseif (strlen($_POST['plz']) < 4) {
-                    return t('Die Postleitzahl muss aus 5 Ziffern bestehen.');
-                }
-            }
-            // Means no error
-            return false;
-        break;
-    }
-}
-
-/**
- * @param string $key
- * @return int
- */
-function Optional($key)
-{
-    global $cfg;
-
-    if ($cfg["signon_show_".$key] <= 1) {
-        return 1;
-    } else {
-        return 0;
-    }
-}
-
-/**
- * @param string $key
- * @return int
- */
-function FieldNeeded($key)
-{
-    global $cfg;
-
-    if ($cfg["signon_show_".$key] == 2) {
-        return 1;
-    } else {
-        return 0;
-    }
-}
-
-/**
- * @param string $key
- * @return int
- */
-function ShowFieldUsrMgr($key)
-{
-    global $cfg;
-
-    if ($cfg["signon_show_".$key] > 0) {
-        return 1;
-    } else {
-        return 0;
-    }
-}
 
 if (!($_GET['mod'] == 'signon' && $auth['login'] && $_GET['party_id'])) {
     $party_user = $db->qry_first("SELECT * FROM %prefix%party_user WHERE user_id = %int% AND party_id= %int%", $_GET["userid"], $party->party_id);
@@ -394,9 +46,14 @@ if (!($_GET['mod'] == 'signon' && $auth['login'] && $_GET['party_id'])) {
   
                     // Module-Permissions
                     $selections = array();
-                    $res = $db->qry("SELECT module.name, module.caption FROM %prefix%modules AS module
+                    $res = $db->qry("
+                      SELECT
+                        module.name,
+                        module.caption
+                      FROM %prefix%modules AS module
                       LEFT JOIN %prefix%menu AS menu ON menu.module = module.name
-                      WHERE menu.file != ''
+                      WHERE
+                        menu.file != ''
                       GROUP BY menu.module");
                     while ($row = $db->fetch_array($res)) {
                           $selections[$row['name']] = $row['caption'];
@@ -412,13 +69,13 @@ if (!($_GET['mod'] == 'signon' && $auth['login'] && $_GET['party_id'])) {
                     }
 
                     $mf->AddField(
-                      t('Zugriffsberechtigung').HTML_NEWLINE.HTML_NEWLINE.
-                      '('.t('Der Benutzertyp muss zusätzlich Admin, oder Superadmin sein.') .')'.HTML_NEWLINE.HTML_NEWLINE.
-                      '('.t('Solange kein Admim einem Modul zugeordnet ist, hat dort jeder Admin Berechtigungen.') .')',
-                      'permissions',
-                      \LanSuite\MasterForm::IS_MULTI_SELECTION,
-                      $selections,
-                      \LanSuite\MasterForm::FIELD_OPTIONAL
+                        t('Zugriffsberechtigung').HTML_NEWLINE.HTML_NEWLINE.
+                        '('.t('Der Benutzertyp muss zusätzlich Admin, oder Superadmin sein.') .')'.HTML_NEWLINE.HTML_NEWLINE.
+                        '('.t('Solange kein Admim einem Modul zugeordnet ist, hat dort jeder Admin Berechtigungen.') .')',
+                        'permissions',
+                        \LanSuite\MasterForm::IS_MULTI_SELECTION,
+                        $selections,
+                        \LanSuite\MasterForm::FIELD_OPTIONAL
                     );
 
                     $mf->AddDropDownFromTable(t('Gruppe'), 'group_id', 'group_id', 'group_name', 'party_usergroups', t('Keine'));
@@ -433,7 +90,7 @@ if (!($_GET['mod'] == 'signon' && $auth['login'] && $_GET['party_id'])) {
             }
   
             $mf->AddField(t('E-Mail'), 'email', '', '', '', CheckValidEmail);
-            $mf->AddField(t('E-Mail wiederholen'),'email2','','','');	
+            $mf->AddField(t('E-Mail wiederholen'), 'email2', '', '', '');
             if (($_GET['action'] != 'change' && $_GET['action'] != 'entrance') || ($_GET['action'] == 'entrance' && !$_GET['userid'])) {
                 if ($cfg['signon_autopw']) {
                     $_SESSION['tmp_pass'] = $usrmgr->GeneratePassword();
@@ -460,12 +117,20 @@ if (!($_GET['mod'] == 'signon' && $auth['login'] && $_GET['party_id'])) {
                 $selections = array();
                 $selections[''] = '---';
                 $PWClans = array();
-                $clans_query = $db->qry("SELECT c.clanid, c.name, c.url, c.password, COUNT(u.clanid) AS members
-                    FROM %prefix%clan AS c
-                    LEFT JOIN %prefix%user AS u ON c.clanid = u.clanid
-                    WHERE u.clanid IS NULL or u.type >= 1
-                    GROUP BY c.clanid
-                    ORDER BY c.name");
+                $clans_query = $db->qry("
+                  SELECT
+                    c.clanid,
+                    c.name,
+                    c.url,
+                    c.password,
+                    COUNT(u.clanid) AS members
+                  FROM %prefix%clan AS c
+                  LEFT JOIN %prefix%user AS u ON c.clanid = u.clanid
+                  WHERE
+                    u.clanid IS NULL
+                    OR u.type >= 1
+                  GROUP BY c.clanid
+                  ORDER BY c.name");
                 while ($row = $db->fetch_array($clans_query)) {
                     $selections[$row['clanid']] = $row['name'] .' ('. $row['members'] .')';
                     if ($row['password']) {
