+++ modules/usrmgr/checkin_assistant.php
@@ -9,7 +9,7 @@ $timestamp    = time();
 if (!$party->party_id) {
     $func->information(t('Es gibt keine aktive Party. Bitte setze im Partymanager eine Party aktiv'));
 } else {
-  // Main-Switch
+    // Main-Switch
     switch ($_GET["step"]) {
         // Auswahl: Angemeldet? ja/Nein
         case '':
@@ -74,15 +74,28 @@ if (!$party->party_id) {
 
             $dsp->NewContent(t('Benutzer hinzufügen'), t('Um einen Benutzer hinzuzufügen, fülle bitte das folgende Formular vollständig aus.'));
 
-            $row = $db->qry_first('SELECT pp.price, pp.price_text FROM %prefix%partys AS p
-        LEFT JOIN %prefix%party_prices AS pp ON p.evening_price_id = pp.price_id
-        WHERE p.party_id = %int%', $party->party_id);
-            $row2 = $db->qry_first('SELECT paid, price, price_text FROM %prefix%party_user AS u
-        LEFT JOIN %prefix%party_prices AS p ON u.price_id = p.price_id
-        WHERE u.party_id = %int% AND u.user_id = %int%', $party->party_id, $_GET['userid']);
+            $row = $db->qry_first('
+              SELECT
+                pp.price,
+                pp.price_text
+              FROM %prefix%partys AS p
+              LEFT JOIN %prefix%party_prices AS pp ON p.evening_price_id = pp.price_id
+              WHERE
+                p.party_id = %int%', $party->party_id);
+            
+            $row2 = $db->qry_first('
+              SELECT
+                paid,
+                price,
+                price_text
+              FROM %prefix%party_user AS u
+              LEFT JOIN %prefix%party_prices AS p ON u.price_id = p.price_id
+              WHERE
+                u.party_id = %int%
+                AND u.user_id = %int%', $party->party_id, $_GET['userid']);
             if ($_GET['mf_step'] != 2 and !$row2['paid']) {
-                    ($row2['price_text'])? $his_wish = '<br>'. t('Der von ihm gewünschte Preis war: '). $row2['price_text'] .' ('. $row2['price'] .' '. $cfg['sys_currency'] .')' : $his_wish = '';
-                    $func->information(t('Achtung: Der Benutzer wird mit der nächsten Seite auf Bezahlt gesetzt.') .'<br>'. t('Preis: ') . $row['price_text'] .' ('. $row['price'] .' '. $cfg['sys_currency'] .')'. $his_wish, NO_LINK);
+                ($row2['price_text'])? $his_wish = '<br>'. t('Der von ihm gewünschte Preis war: '). $row2['price_text'] .' ('. $row2['price'] .' '. $cfg['sys_currency'] .')' : $his_wish = '';
+                $func->information(t('Achtung: Der Benutzer wird mit der nächsten Seite auf Bezahlt gesetzt.') .'<br>'. t('Preis: ') . $row['price_text'] .' ('. $row['price'] .' '. $cfg['sys_currency'] .')'. $his_wish, NO_LINK);
             }
 
             include_once("modules/usrmgr/add.php");
@@ -95,19 +108,26 @@ if (!$party->party_id) {
                 $row = $db->qry_first('SELECT evening_price_id FROM %prefix%partys WHERE party_id = %int%', $party->party_id);
         
                 if ($row2['paid']) {
-                    $db->qry('UPDATE %prefix%party_user SET
-            checkin = NOW()
-            WHERE user_id = %int% AND party_id = %int%', $_GET['userid'], $party->party_id);
+                    $db->qry('
+                      UPDATE %prefix%party_user
+                      SET
+                        checkin = NOW()
+                      WHERE
+                        user_id = %int%
+                        AND party_id = %int%', $_GET['userid'], $party->party_id);
                 } else {
                     $db->qry('DELETE FROM %prefix%party_user WHERE user_id = %int% AND party_id = %int%', $_GET['userid'], $party->party_id);
-                    $db->qry(
-                        'INSERT INTO %prefix%party_user SET
-            user_id = %int%, party_id = %int%, price_id = %int%, checkin = NOW(),
-            paid = 2, paiddate = NOW(), seatcontrol = 0, signondate = NOW()',
-                        $_GET['userid'],
-                        $party->party_id,
-                        $row['evening_price_id']
-                    );
+                    $db->qry('
+                      INSERT INTO %prefix%party_user
+                      SET
+                        user_id = %int%,
+                        party_id = %int%,
+                        price_id = %int%,
+                        checkin = NOW(),
+                        paid = 2,
+                        paiddate = NOW(),
+                        seatcontrol = 0,
+                        signondate = NOW()', $_GET['userid'], $party->party_id, $row['evening_price_id']);
                 }
             }
             break;
@@ -151,9 +171,19 @@ if (!$party->party_id) {
 
         // Belegten Sitzplatz tauschen / löschen?
         case 9:
-            $seat = $db->qry_first("SELECT s.userid, s.status, u.username, u.firstname, u.name FROM %prefix%seat_seats AS s
-  			LEFT JOIN %prefix%user AS u ON s.userid = u.userid
-  			WHERE blockid = %int% AND row = %string% AND col = %string%", $_GET['blockid'], $_GET['row'], $_GET['col']);
+            $seat = $db->qry_first("
+              SELECT
+                s.userid,
+                s.status,
+                u.username,
+                u.firstname,
+                u.name
+              FROM %prefix%seat_seats AS s
+              LEFT JOIN %prefix%user AS u ON s.userid = u.userid
+              WHERE
+                blockid = %int%
+                AND row = %string%
+                AND col = %string%", $_GET['blockid'], $_GET['row'], $_GET['col']);
 
             if ($seat['status'] == 1 or $seat['status'] == 3) {
                 $_GET['step'] = 10;
