+++ modules/usrmgr/party.php
@@ -1,91 +1,28 @@
 <?php
 
-use LanSuite\Module\Seating\Seat2;
+$mail = new \LanSuite\Module\Mail\Mail();
+$usrmgr = new \LanSuite\Module\UsrMgr\UserManager($mail);
 
-include_once("modules/usrmgr/class_usrmgr.php");
-$usrmgr = new UsrMgr();
-
-$seat2 = new Seat2();
-
-/**
- * @return bool
- */
-function PartyMail()
-{
-    global $usrmgr, $func, $mail, $auth;
-
-    $usrmgr->WriteXMLStatFile();
-
-    if ($_POST['sendmail'] or $auth['type'] < 2) {
-        if ($usrmgr->SendSignonMail(1)) {
-            $func->confirmation(t('Eine Bestätigung der Anmeldung wurde an deine E-Mail-Adresse gesendet.'), NO_LINK);
-        } else {
-            $func->error(t('Es ist ein Fehler beim Versand der Informations-E-Mail aufgetreten.'). $mail->error, NO_LINK);
-        }
-    }
-
-    return true;
-}
+$seat2 = new \LanSuite\Module\Seating\Seat2();
 
 if ($party->count == 0) {
     $func->information(t('Aktuell sind keine Partys geplant.'), 'index.php?mod='. $_GET['mod']);
 } else {
     if ($_GET['user_id'] == $auth['userid'] or $auth['type'] >= 2) {
-        function ChangeAllowed($id)
-        {
-            global $db, $row, $func, $auth, $seat2;
-
-            // Do not allow changes, if party is over
-            if ($row['enddate'] < time()) {
-                return t('Du kannst dich nicht mehr zu dieser Party an-, oder abmelden, da sie bereits vorüber ist');
-            }
-
-            // Signon started?
-            if ($row['sstartdate'] > time()) {
-                return t('Die Anmeldung öffnet am'). HTML_NEWLINE .'<strong>'. $func->unixstamp2date($row['sstartdate'], 'daydatetime'). '</strong>';
-            }
-
-            // Signon ended?
-            if ($row['senddate'] < time() and $auth['type'] < 2) {
-                return t('Die Anmeldung ist beendet seit'). HTML_NEWLINE .'<strong>'. $func->unixstamp2date($row['senddate'], 'daydatetime'). '</strong>';
-            }
-
-            // Do not allow changes, if user has paid
-            if ($auth['type'] <= 1) {
-                $row2 = $db->qry_first("SELECT paid FROM %prefix%party_user WHERE party_id = %int% AND user_id = %int%", $_GET['party_id'], $id);
-                if ($row2['paid']!= 0) {
-                    return t('Du bist für diese Party bereits auf bezahlt gesetzt. Bitte einen Admin dich auf "nicht bezahlt" zu setzen, bevor du dich abmeldest');
-                }
-            }
-
-            // Check age
-            if (isset($_POST['InsertControll1']) && $_POST['InsertControll1']) {
-                $res = $db->qry("SELECT %prefix%partys.minage FROM %prefix%user, %prefix%partys
-                            WHERE %prefix%partys.party_id = %int%
-                                AND %prefix%user.userid = %int%
-                                AND DATEDIFF(DATE_SUB(%prefix%partys.startdate, INTERVAL %prefix%partys.minage YEAR), %prefix%user.birthday) < 0
-                                AND %prefix%partys.minage > 0", $_GET['party_id'], $id);
-                      $minage = $db->fetch_array($res);
-                      $db->free_result($res);
-                if (isset($minage['minage'])) {
-                    return t('Du must mindestens %1 Jahre alt sein um an dieser Party teilnehmen zu d&uuml;rfen!', $minage['minage']);
-                }
-            }
-
-            $row2 = $db->qry_first("SELECT paid FROM %prefix%party_user WHERE party_id = %int% AND user_id = %int%", $_GET['party_id'], $id);
-
-            // Free seats if the user hasn't paid already
-            if ($row2['paid'] == 0) {
-                $seat2->FreeSeatAllMarkedByUser($id);
-            }
-
-            return false;
-        }
-
         // Show Upcomming
         $MFID = 1;
 
-        $res = $db->qry("SELECT *, UNIX_TIMESTAMP(enddate) AS enddate, UNIX_TIMESTAMP(sstartdate) AS sstartdate, UNIX_TIMESTAMP(senddate) AS senddate, UNIX_TIMESTAMP(startdate) AS startdate FROM %prefix%partys WHERE UNIX_TIMESTAMP(enddate) >= UNIX_TIMESTAMP(NOW()) ORDER BY startdate");
+        $res = $db->qry("
+          SELECT
+            *,
+            UNIX_TIMESTAMP(enddate) AS enddate,
+            UNIX_TIMESTAMP(sstartdate) AS sstartdate,
+            UNIX_TIMESTAMP(senddate) AS senddate,
+            UNIX_TIMESTAMP(startdate) AS startdate
+          FROM %prefix%partys
+          WHERE
+            UNIX_TIMESTAMP(enddate) >= UNIX_TIMESTAMP(NOW())
+          ORDER BY startdate");
         while ($row = $db->fetch_array($res)) {
             if ($_GET['mf_step'] != 2 or $row['party_id'] == $_GET['party_id']) {
                 $dsp->AddFieldsetStart($row['name'] .' ('. $func->unixstamp2date($row['startdate'], 'datetime') .' - '. $func->unixstamp2date($row['enddate'], 'datetime') .')');
@@ -151,20 +88,26 @@ if ($party->count == 0) {
 
         // ShowHistory
         $dsp->AddFieldsetStart(t('Vergangene Partys'));
-        $res = $db->qry("SELECT
-          p.*
-        , pu.user_id
-        , pu.paid
-        , UNIX_TIMESTAMP(pu.checkin) AS checkin
-        , UNIX_TIMESTAMP(pu.checkout) AS checkout
-        , UNIX_TIMESTAMP(p.enddate) AS enddate
-        , UNIX_TIMESTAMP(p.sstartdate) AS sstartdate
-        , UNIX_TIMESTAMP(p.senddate) AS senddate
-        , UNIX_TIMESTAMP(p.startdate) AS startdate
-      FROM %prefix%partys AS p
-      LEFT JOIN %prefix%party_user AS pu ON p.party_id = pu.party_id
-      WHERE UNIX_TIMESTAMP(p.enddate) < UNIX_TIMESTAMP(NOW()) AND (pu.user_id = %int% OR pu.user_id IS NULL)
-      ORDER BY p.startdate", $_GET['user_id']);
+        $res = $db->qry("
+          SELECT
+            p.*,
+            pu.user_id,
+            pu.paid,
+            UNIX_TIMESTAMP(pu.checkin) AS checkin,
+            UNIX_TIMESTAMP(pu.checkout) AS checkout,
+            UNIX_TIMESTAMP(p.enddate) AS enddate,
+            UNIX_TIMESTAMP(p.sstartdate) AS sstartdate,
+            UNIX_TIMESTAMP(p.senddate) AS senddate,
+            UNIX_TIMESTAMP(p.startdate) AS startdate
+          FROM %prefix%partys AS p
+          LEFT JOIN %prefix%party_user AS pu ON p.party_id = pu.party_id
+          WHERE
+            UNIX_TIMESTAMP(p.enddate) < UNIX_TIMESTAMP(NOW())
+            AND (
+              pu.user_id = %int%
+              OR pu.user_id IS NULL
+            )
+          ORDER BY p.startdate", $_GET['user_id']);
         while ($row = $db->fetch_array($res)) {
             $text = '';
             if ($row['user_id']) {
