+++ modules/usrmgr/search.inc.php
@@ -1,87 +1,6 @@
 <?php
+
 include_once('modules/usrmgr/search_main.inc.php');
-
-/**
- * @param string $clan_name
- * @return string
- */
-function ClanURLLinkUsrMgrSearchInc($clan_name)
-{
-    global $line, $func;
-
-    if ($clan_name == '') {
-        return '';
-    } elseif ($func->isModActive('clanmgr')) {
-        return '<a href="index.php?mod=clanmgr&action=clanmgr&step=2&clanid='. $line['clanid'] .'">'. $clan_name .'</a>';
-    } elseif ($clan_name != '' and $line['clanurl'] != '' and $line['clanurl'] != 'http://') {
-        if (substr($line['clanurl'], 0, 7) != 'http://') {
-            $line['clanurl'] = 'http://'. $line['clanurl'];
-        }
-        return '<a href="'. $line['clanurl'] .'" target="_blank">'. $clan_name .'</a>';
-    } else {
-        return $clan_name;
-    }
-}
-
-/**
- * @param int $userid
- * @return bool
- */
-function IfLowerUserlevel($userid)
-{
-    global $line, $auth;
-  
-    if ($line['type'] < $auth['type']) {
-        return true;
-    } else {
-        return false;
-    }
-}
-
-/**
- * @param int $userid
- * @return bool
- */
-function IfLowerOrEqualUserlevel($userid)
-{
-    global $line, $auth;
-
-    if ($line['type'] <= $auth['type']) {
-        return true;
-    } else {
-        return false;
-    }
-}
-
-/**
- * @param int $userid
- * @return bool
- */
-function IfLocked($userid)
-{
-    global $line;
-
-    if ($line['locked']) {
-        return true;
-    } else {
-        return false;
-    }
-}
-
-/**
- * @param int $userid
- * @return bool
- */
-function IfUnlocked($userid)
-{
-    global $line;
-
-    if (!$line['locked']) {
-        return true;
-    } else {
-        return false;
-    }
-}
 
 $ms2->AddTextSearchField(t('Clan'), array('c.name' => 'like'));
