+++ modules/usrmgr/settings.php
@@ -2,26 +2,6 @@
 
 $gd = new \LanSuite\GD();
 
-/**
- * @param string $AvatarName
- * @return bool|string
- */
-function CheckAndResizeUploadPic($AvatarName)
-{
-    global $gd;
-
-    if ($AvatarName == '') {
-        return false;
-    }
-    $FileEnding = strtolower(substr($AvatarName, strrpos($AvatarName, '.'), 5));
-    if ($FileEnding != '.png' and $FileEnding != '.gif' and $FileEnding != '.jpg' and $FileEnding != '.jpeg') {
-        return t('Bitte eine Grafikdatei auswählen');
-    }
-
-    $gd->CreateThumb($AvatarName, $AvatarName, 100, 100);
-    return false;
-}
-
 $mf = new \LanSuite\MasterForm();
 
 // Designs
