+++ modules/usrmgr/user_fields.php
@@ -1,31 +1,5 @@
 <?php
 
-/**
- * @param int $id
- * @return bool
- */
-function UpdateUsrMgrUserFields($id)
-{
-    global $db;
-  
-    $db->qry("ALTER TABLE %prefix%user ADD %plain% VARCHAR(255) NOT NULL;", $_POST['name']);
-  
-    return true;
-}
-
-/**
- * @param string $val
- * @return bool|string
- */
-function check_no_space($val)
-{
-    if (strpos($val, ' ') !== false) {
-        return t('Der Feldname darf kein Leerzeichen enthalten');
-    } else {
-        return false;
-    }
-}
-
 switch ($_GET['step']) {
     default:
         $ms2 = new \LanSuite\Module\MasterSearch2\MasterSearch2('usrmgr');
